feat(merchant-sdk): Added / modified some tests
EC-91
